rename package 'delete' to 'mockobjects'
